+++ examples/min-android-sample/app/src/main/java/io/realm/example/minandroidsample/android/MainActivity.kt
@@ -1,8 +1,8 @@
 package io.realm.example.minandroidsample.android
 
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
 import io.realm.sample.minandroidsample.Greeting
 
 class MainActivity : AppCompatActivity() {