+++ packages/plugin-compiler/src/test/kotlin/io/realm/kotlin/compiler/GenerationExtensionTest.kt
@@ -377,8 +377,10 @@ class GenerationExtensionTest {
             messageOutputStream = System.out
             componentRegistrars = plugins
             inheritClassPath = true
+            jvmTarget = "17"
             kotlincArguments = listOf(
                 "-Xjvm-default=all-compatibility",
+                "-Xverify-ir=error",
                 "-Xdump-directory=${inputs.outputDir()}",
                 "-Xphases-to-dump-after=JvmValidateIrBeforeLowering"
             )